[fix] 초대 회원가입 로직 수정
